+++ configupdate.py
@@ -127,7 +127,7 @@ if "0.6.2 beta" in versioninfo2:
     except configparser.DuplicateSectionError:
         print("Failed to add the hurricane section. Does it exist?")
 
-    print("14 new configuration options have been added from 0.6.2 beta to 0.6.3 beta.",
+    print("17 new configuration options have been added from 0.6.2 beta to 0.6.3 beta.",
           "Details:",
           "- FIRSTINPUT/geoipservice_enabled - Sets if the service to allow current location queries is enabled - Defaults to False",
           "- FIRSTINPUT/allow_pwsqueries - Sets if PyWeather will allow PWS queries - Defaults to True",
@@ -142,7 +142,10 @@ if "0.6.2 beta" in versioninfo2:
           "- FAVORITE LOCATIONS/favloc4 - Sets the fourth favorite location - Defaults to 'None'",
           "- FAVORITE LOCATIONS/favloc5 - Sets the fifth favorite location - Defaults to 'None'",
           "- GEOCODER API/customkey_enabled - Sets if a custom API key for the geocoder is enabled - Defaults to 'False'.",
-          "- GEOCODER API/customkey - Sets the custom API key for the geocoder, if enabled - Defaults to 'None'.", sep="\n")
+          "- GEOCODER API/customkey - Sets the custom API key for the geocoder, if enabled - Defaults to 'None'.",
+          "- SUMMARY/showyesterdayonsummary - Sets if yesterday's weather data should be shown on the summary screen - Defaults to 'False'.",
+          "- PREFETCH/yesterdaydata_atboot - Sets if yesterday's weather data should be prefetched at boot. - Defaults to 'False'.",
+          "- CACHE/yesterday_cachedtime - Sets the cache time for yesterday's weather - Defaults to '720'.", sep="\n")
 
     print("", "One old configuration option has been removed - Please delete this options from your config file.",
           "- UPDATER/allowGitForUpdating - Used to set if the Git updater was enabled - Now removed", sep="\n")
@@ -161,6 +164,9 @@ if "0.6.2 beta" in versioninfo2:
     config['FAVORITE LOCATIONS']['favloc5'] = 'None'
     config['GEOCODER API']['customkey_enabled'] = 'False'
     config['GEOCODER API']['customkey'] = 'None'
+    config['SUMMARY']['showyesterdayonsummary'] = 'False'
+    config['PREFETCH']['yesterdaydata_atboot'] = 'False'
+    config['CACHE']['yesterday_cachedtime'] = '720'
 
 elif "0.6.1 beta" in versioninfo2:
     try:
@@ -193,7 +199,7 @@ elif "0.6.1 beta" in versioninfo2:
     except configparser.DuplicateSectionError:
         print("Failed to add the hurricane section. Does it exist?")
 
-    print("20 new configuration options have been added from 0.6.1 beta to 0.6.3 beta.",
+    print("23 new configuration options have been added from 0.6.1 beta to 0.6.3 beta.",
           "Details:",
           "- CACHE/tide_cachedtime - Sets the cache time on tide data - Defaults to 480",
           "- SUMMARY/showtideonsummary - Sets if tide data should be shown on the summary screen - Defaults to False",
@@ -216,7 +222,10 @@ elif "0.6.1 beta" in versioninfo2:
           "- FAVORITE LOCATIONS/favloc4 - Sets the fourth favorite location - Defaults to 'None'",
           "- FAVORITE LOCATIONS/favloc5 - Sets the fifth favorite location - Defaults to 'None'",
           "- GEOCODER API/customkey_enabled - Sets if a custom API key for the geocoder is enabled - Defaults to 'False'.",
-          "- GEOCODER API/customkey - Sets the custom API key for the geocoder, if enabled - Defaults to 'None'.", sep="\n")
+          "- GEOCODER API/customkey - Sets the custom API key for the geocoder, if enabled - Defaults to 'None'.",
+          "- SUMMARY/showyesterdayonsummary - Sets if yesterday's weather data should be shown on the summary screen - Defaults to 'False'.",
+          "- PREFETCH/yesterdaydata_atboot - Sets if yesterday's weather data should be prefetched at boot. - Defaults to 'False'.",
+          "- CACHE/yesterday_cachedtime - Sets the cache time for yesterday's weather - Defaults to '720'.", sep="\n")
 
     print("")
     print("3 old configuration options, and 2 sections have been removed. Please delete these options from your config file.",
@@ -248,6 +257,9 @@ elif "0.6.1 beta" in versioninfo2:
     config['FAVORITE LOCATIONS']['favloc5'] = 'None'
     config['GEOCODER API']['customkey_enabled'] = 'False'
     config['GEOCODER API']['customkey'] = 'None'
+    config['SUMMARY']['showyesterdayonsummary'] = 'False'
+    config['PREFETCH']['yesterdaydata_atboot'] = 'False'
+    config['CACHE']['yesterday_cachedtime'] = '720'
     geopycheck()
 elif "0.6 beta" or "0.6.0.1 beta" in versioninfo2:
     # A usual input() and sys.exit() isn't present here, as it's assumed this
@@ -292,7 +304,7 @@ elif "0.6 beta" or "0.6.0.1 beta" in versioninfo2:
     except configparser.DuplicateSectionError:
         print("Failed to add the hurricane section. Does it exist?")
         
-    print("28 new configuration options have been added from 0.6 beta to 0.6.3 beta.",
+    print("31 new configuration options have been added from 0.6 beta to 0.6.3 beta.",
           "Details:",
           "- CACHE/alerts_cachedtime - Sets the cache time on alert data - Defaults to 5",
           "- CACHE/current_cachedtime - Sets the cache time on current data - Defaults to 10",
@@ -323,7 +335,10 @@ elif "0.6 beta" or "0.6.0.1 beta" in versioninfo2:
           "- FAVORITE LOCATIONS/favloc4 - Sets the fourth favorite location - Defaults to 'None'",
           "- FAVORITE LOCATIONS/favloc5 - Sets the fifth favorite location - Defaults to 'None'",
           "- GEOCODER API/customkey_enabled - Sets if a custom API key for the geocoder is enabled - Defaults to 'False'.",
-          "- GEOCODER API/customkey - Sets the custom API key for the geocoder, if enabled - Defaults to 'None'.", sep="\n")
+          "- GEOCODER API/customkey - Sets the custom API key for the geocoder, if enabled - Defaults to 'None'.",
+          "- SUMMARY/showyesterdayonsummary - Sets if yesterday's weather data should be shown on the summary screen - Defaults to 'False'.",
+          "- PREFETCH/yesterdaydata_atboot - Sets if yesterday's weather data should be prefetched at boot. - Defaults to 'False'.",
+          "- CACHE/yesterday_cachedtime - Sets the cache time for yesterday's weather - Defaults to '720'.", sep="\n")
 
     print("")
     print("3 old configuration options, and 2 sections are now removed. Please delete these options from your config file.",
@@ -363,6 +378,9 @@ elif "0.6 beta" or "0.6.0.1 beta" in versioninfo2:
     config['FAVORITE LOCATIONS']['favloc5'] = 'None'
     config['GEOCODER API']['customkey_enabled'] = 'False'
     config['GEOCODER API']['customkey'] = 'None'
+    config['SUMMARY']['showyesterdayonsummary'] = 'False'
+    config['PREFETCH']['yesterdaydata_atboot'] = 'False'
+    config['CACHE']['yesterday_cachedtime'] = '720'
     geopycheck()
 else:
     print("Hmm. Your version identifier didn't match any known versions.",